pyproject.toml
[tool.poetry]
name = "libertai-backend"
version = "0.0.1"
description = ""
authors = ["ALI EL BROUDI"]
readme = "README.md"
homepage = "https://libertai.io"
repository = "https://github.com/LibertAI/libertai-backend"
documentation = "https://docs.libertai.io"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = { extras = ["standard"], version = "^0.114.1" }
aleph-sdk-python = "^1.0.1"
web3 = "^6.3.0"
libertai-utils = {git = "git@github.com:Libertai/libertai-utils.git", rev = "main"}

[tool.poetry.group.dev.dependencies]
mypy = "^1.11.1"
ruff = "^0.6.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
